# restaurant/src/main/java/hu/oe/yokudlela/restaurant/generated/rest/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.oe.yokudlela.restaurant.generated.rest.api;

import hu.oe.yokudlela.restaurant.generated.rest.model.ApiError;
import hu.oe.yokudlela.restaurant.generated.rest.model.MenuIdGet400Response;
import hu.oe.yokudlela.restaurant.generated.rest.model.MenuItem;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    /**
     * GET /menu : Menüelemek listázása
     * Lekérdezi az összes ételt és italt.
     *
     * @return Menüelemek sikeresen lekérdezve (status code 200)
     */
    @Operation(
        operationId = "menuGet",
        summary = "Menüelemek listázása",
        description = "Lekérdezi az összes ételt és italt.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Menüelemek sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MenuItem.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<MenuItem>> menuGet(
        
    );


    /**
     * GET /menu/gluten-free : List only gluten-free menu items
     *
     * @return Successfully retrieved gluten-free menu items (status code 200)
     *         or Bad request (status code 400)
     */
    @Operation(
        operationId = "menuGlutenFreeGet",
        summary = "List only gluten-free menu items",
        tags = { "Default" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved gluten-free menu items", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MenuItem.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu/gluten-free",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<MenuItem>> menuGlutenFreeGet(
        
    );


    /**
     * DELETE /menu/{id} : Menüelem törlése
     * Menüelem törlése ID alapján.
     *
     * @param id  (required)
     * @return Menüelem sikeresen törölve (status code 204)
     *         or A menüelem nem található (status code 400)
     */
    @Operation(
        operationId = "menuIdDelete",
        summary = "Menüelem törlése",
        description = "Menüelem törlése ID alapján.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Menüelem sikeresen törölve"),
            @ApiResponse(responseCode = "400", description = "A menüelem nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuIdGet400Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/menu/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> menuIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * GET /menu/{id} : Menüelem lekérdezése
     * Egy adott menüelem lekérdezése ID alapján.
     *
     * @param id  (required)
     * @return Menüelem sikeresen lekérdezve (status code 200)
     *         or A menüelem nem található (status code 400)
     */
    @Operation(
        operationId = "menuIdGet",
        summary = "Menüelem lekérdezése",
        description = "Egy adott menüelem lekérdezése ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Menüelem sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuItem.class))
            }),
            @ApiResponse(responseCode = "400", description = "A menüelem nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuIdGet400Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/menu/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<MenuItem> menuIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * PUT /menu/{id} : Menüelem módosítása
     * Menüelem módosítása ID alapján.
     *
     * @param id  (required)
     * @param menuItem  (required)
     * @return Menüelem sikeresen módosítva (status code 200)
     *         or A menüelem nem található (status code 400)
     */
    @Operation(
        operationId = "menuIdPut",
        summary = "Menüelem módosítása",
        description = "Menüelem módosítása ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Menüelem sikeresen módosítva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuItem.class))
            }),
            @ApiResponse(responseCode = "400", description = "A menüelem nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuIdGet400Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/menu/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<MenuItem> menuIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "MenuItem", description = "", required = true) @Valid @RequestBody MenuItem menuItem
    );


    /**
     * POST /menu : Új menüelem hozzáadása
     * Új étel vagy ital hozzáadása a menühöz.
     *
     * @param menuItem  (required)
     * @return Menüelem sikeresen hozzáadva (status code 201)
     *         or A menüelem már létezik (status code 400)
     */
    @Operation(
        operationId = "menuPost",
        summary = "Új menüelem hozzáadása",
        description = "Új étel vagy ital hozzáadása a menühöz.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Menüelem sikeresen hozzáadva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MenuItem.class))
            }),
            @ApiResponse(responseCode = "400", description = "A menüelem már létezik", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/menu",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<MenuItem> menuPost(
        @Parameter(name = "MenuItem", description = "", required = true) @Valid @RequestBody MenuItem menuItem
    );

}