# restaurant/target/classes/openapi/restaurant.yaml
openapi: 3.0.0
info:
  title: Restaurant Menu API
  description: API a restaurant menu items nyilvántartására.
  version: 1.0.0

servers:
  - url: https://oe-html-5-2025-tavasz.wiremockapi.cloud/api
  - url: http://localhost:8080/api
    description: Helyi fejlesztői szerver


paths:
  /menu/gluten-free:
    get:
      summary: List only gluten-free menu items
      operationId: menuGlutenFreeGet
      tags:
        - Default
      responses:
        '200':
          description: Successfully retrieved gluten-free menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /menu:
    get:
      summary: Menüelemek listázása
      description: Lekérdezi az összes ételt és italt.
      responses:
        "200":
          description: Menüelemek sikeresen lekérdezve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
              example:
                - id: 1
                  name: Pizza Margherita
                  category: MAIN_COURSE
                  sugarFree: false
                  glutenFree: false
                - id: 2
                  name: Cheesecake
                  category: DESSERT
                  sugarFree: false
                  glutenFree: false
                - id: 3
                  name: Cola Zero
                  category: DRINK
                  sugarFree: true
                  glutenFree: true

    post:
      summary: Új menüelem hozzáadása
      description: Új étel vagy ital hozzáadása a menühöz.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItem"
      responses:
        "201":
          description: Menüelem sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: A menüelem már létezik
          content:
            application/json:
              schema:
                type: object
                schema:
                  $ref: "#/components/schemas/ApiError"
                example:
                  path: "/restaurant"
                  message: "error.validation"
                  errors:
                    - "error.menuitem.name.exists"
  paths:
    /menu/gluten-free:
      get:
        summary: List only gluten-free menu items
        operationId: menuGlutenFreeGet
        tags:
          - Default
        responses:
          '200':
            description: Successfully retrieved gluten-free menu items
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/MenuItem'
          '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ApiError'
  /menu/{id}:
    get:
      summary: Menüelem lekérdezése
      description: Egy adott menüelem lekérdezése ID alapján.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menüelem sikeresen lekérdezve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: A menüelem nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Menüelem nem található

    put:
      summary: Menüelem módosítása
      description: Menüelem módosítása ID alapján.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItem"
      responses:
        "200":
          description: Menüelem sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: A menüelem nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Menüelem nem található

    delete:
      summary: Menüelem törlése
      description: Menüelem törlése ID alapján.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Menüelem sikeresen törölve
        "400":
          description: A menüelem nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Menüelem nem található

components:
  schemas:
    MenuItem:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: integer
          description: Menüelem azonosítója

        name:
          type: string
          description: Az étel vagy ital neve
          x-field-extra-annotation: '@hu.oe.yokudlela.restaurant.validation.NameExists(message="error.menuitem.name.exists")'

        category:
          type: string
          enum:
            - MAIN_COURSE
            - DESSERT
            - DRINK

        sugarFree:
          type: boolean
          description: Cukormentes-e

        glutenFree:
          type: boolean
          description: Gluténmentes-e

    ApiError:
      type: object
      properties:
        path:
          type: string
          description: A kérés útvonala (pl. /menu)

        message:
          type: string
          description: Fő hibaüzenet (pl. error.validation)

        errors:
          type: array
          items:
            type: string
          description: Részletes validációs hibák listája